# digiwf-integrations/digiwf-oce-cosmos-integration/digiwf-oce-cosmos-integration-client/apidoc/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "The sysHUB REST API / SDK v3 is a RESTful API that provides third party teams access to sysHUB features through a standard way. This page documents all operations intended for third-party integration.",
    "version": "v3",
    "title": "sysHUB REST-API"
  },
  "host": "",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/webapi/v3/backuprestore/metadata": {
      "get": {
        "tags": [
          "Backup/Restore operations"
        ],
        "summary": "Gets the Backup properties",
        "description": "Gets the Backup properties and all backup types of a given folder. Requires permissions 'PERM_IEPOSSERVER_RESTOREDATABASE', 'PERM_IEPOSSERVER_EXISTSSERVERFILE'.",
        "operationId": "getBackupMetadataUsingGET",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "description": "The folder to the backup files, file-separator = '/'",
            "required": true,
            "type": "string",
            "maxLength": 256,
            "minLength": 0,
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        },
        "security": [
          {
            "oauth2schema": [
              "public",
              "private"
            ]
          }
        ]
      }
    },
    "/webapi/v3/backuprestore/{command}": {
      "post": {
        "tags": [
          "Backup/Restore operations"
        ],
        "summary": "Executes a Backup or Restore",
        "description": "Executes a Backup or Restore. Requires permissions 'PERM_IEPOSSERVER_SAVEDATABASE', 'PERM_IEPOSSERVER_RESTOREDATABASE'.",
        "operationId": "doBackupRestoreUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "command",
            "in": "path",
            "description": "backup or restore",
            "required": true,
            "type": "string"
          },
          {
            "name": "folder",
            "in": "query",
            "description": "The folder for the operation",
            "required": true,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "in": "body",
            "name": "properties",
            "description": "Representation of the backup properties and which backuptypes should be created/restored",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BackupRestoreDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SimpleResponse"
            }
          },
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SimpleResponse"
            }
          }
        },
        "security": [
          {
            "oauth2schema": [
              "public",
              "private"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "oauth2schema": {
      "type": "oauth2",
      "tokenUrl": "",
      "flow": "password",
      "scopes": {
        "public": "public access",
        "private": "private access only for internal use"
      }
    }
  },
  "definitions": {
    "BackupRestoreDTO": {
      "type": "object",
      "required": [
        "BACKUPTYPES"
      ],
      "properties": {
        "BACKUPDESCRIPTION": {
          "type": "string",
          "example": "Initial Backup for COSMOS 2021",
          "description": "The description of the backup (command = backup)"
        },
        "BACKUPNAME": {
          "type": "string",
          "example": "Initial Backup",
          "description": "The name of the backup (command = backup)"
        },
        "BACKUPTYPES": {
          "type": "array",
          "example": "{[\"USER\", \"CONFIG\"]}",
          "description": "List with backuptypes to restore or which are in a backup folder",
          "items": {
            "type": "string",
            "enum": [
              "CONFIG",
              "CONFIGNOPACKAGE",
              "CONFIGRAW",
              "JOB",
              "JOBLOG",
              "JOBSELECTION",
              "MISC",
              "SYSLOG",
              "USER",
              "USERLOG",
              "USERTRANSPORT"
            ]
          }
        }
      },
      "title": "BackupRestoreDTO",
      "description": "The model for backup and restore information"
    },
...
    "SimpleResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "id",
          "description": "The name of the response object"
        },
        "type": {
          "type": "string",
          "example": "result",
          "description": "The type of the value as an enumeration value",
          "enum": [
            "message",
            "result"
          ]
        },
        "value": {
          "type": "object",
          "example": 12,
          "description": "The value describes the key properties value"
        }
      },
      "title": "SimpleResponse",
      "xml": {
        "name": "simpleResponse",
        "attribute": false,
        "wrapped": false
      },
      "description": "The SimpleResponse object contains only a name and a value and the content is depend to the request"
    },
...
  }
}